pyproject.toml
[project]
name = "ros-mcp-server"
version = "0.2.0"
description = "MCP tool server for ROS using rosbridge WebSocket"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "you@example.com" }
]

dependencies = [
    "mcp[cli]>=1.6.0",
    "opencv-python>=4.11.0.86",
    "websocket-client>=1.8.0"
]

[project.optional-dependencies]
dev = [
    "ruff",    # single tool for linting & formatting
    "pytest"   # for tests
]

# ---------------- Ruff configuration ----------------
[tool.ruff]
line-length = 100
target-version = "py310"

# Linter config moved to new section
[tool.ruff.lint]
select = ["E", "F", "I"]  # basic errors, pyflakes, import sorting
ignore = ["E501"]         # ignore long-line warnings (optional)

# Formatter config stays here
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
docstring-code-format = true
